bytemuck_derive

Module traits

Macros

bail 🔒

mk_repr 🔒

Structs

AnyBitPattern

CheckedBitPattern

Contiguous

NoUninit

Pod

Representation 🔒

TransparentWrapper

VariantDiscriminantIterator 🔒

WrappedType 🔒

Zeroable

Enums

IntegerRepr 🔒

Repr 🔒

Constants

GENERATED_TYPE_DOCUMENTATION 🔒

Traits

Derivable

Functions

bytemuck_crate_name

enum_has_fields 🔒

generate_assert_no_padding 🔒

Check that a struct or enum has no padding by asserting that the size of the type is equal to the sum of the size of it’s fields and discriminant (for enums, this must be asserted for each variant).

generate_checked_bit_pattern_enum 🔒

generate_checked_bit_pattern_enum_with_fields 🔒

generate_checked_bit_pattern_enum_without_fields 🔒

generate_checked_bit_pattern_struct 🔒

generate_enum_discriminant 🔒

generate_fields_are_trait 🔒

Check that all fields implement a given trait

get_enum_discriminant 🔒

Get the type of an enum’s discriminant.

get_enum_variants 🔒

get_field_types 🔒

get_fields 🔒

Extract the Fields off a DeriveInput, or, in the enum case, off those of the enum_variant, when provided (e.g., for Zeroable).

get_repr 🔒

get_struct_fields 🔒

get_type_from_simple_attr 🔒

get a simple #[foo(bar)] attribute, returning bar

get_wrapped_type_from_stream 🔒

get_zero_variant 🔒

Helper function to get the variant with discriminant zero (implicit or explicit).

parse_int_expr 🔒