Crate bytemuck_derive

Derive macros for bytemuck traits.

Modules

traits 🔒

Functions

add_trait_marker 🔒

Add a trait marker to the generics if it is not already present

derive_marker_trait 🔒

Basic wrapper for error handling

derive_marker_trait_inner 🔒

find_and_parse_helper_attributes 🔒

Find #[name(key = "value")] helper attributes on the struct, and return their "value"s parsed with parser.

Derive Macros

AnyBitPattern

Derive the AnyBitPattern trait for a struct

ByteEq

Derive the PartialEq and Eq trait for a type

ByteHash

Derive the Hash trait for a type

CheckedBitPattern

Derive the CheckedBitPattern trait for a struct or enum.

Contiguous

Derive the Contiguous trait for an enum

NoUninit

Derive the NoUninit trait for a struct or enum

Pod

Derive the Pod trait for a struct

TransparentWrapper

Derive the TransparentWrapper trait for a struct

Zeroable

Derive the Zeroable trait for a struct