Module bytemuck_derive::traits

Macros

bail 🔒

mk_repr 🔒

Structs

AnyBitPattern

CheckedBitPattern

Contiguous

NoUninit

Pod

Representation 🔒

TransparentWrapper

VariantDiscriminantIterator 🔒

Zeroable

Enums

IntegerRepr 🔒

Repr 🔒

Traits

Derivable

Functions

bytemuck_crate_name

enum_has_fields 🔒

generate_assert_no_padding 🔒

Check that a struct has no padding by asserting that the size of the struct is equal to the sum of the size of it’s fields

generate_checked_bit_pattern_enum 🔒

generate_checked_bit_pattern_enum_with_fields 🔒

generate_checked_bit_pattern_enum_without_fields 🔒

generate_checked_bit_pattern_struct 🔒

generate_fields_are_trait 🔒

Check that all fields implement a given trait

get_enum_variants 🔒

get_field_types 🔒

get_fields 🔒

get_ident_from_stream 🔒

get_repr 🔒

get_simple_attr 🔒

get a simple #[foo(bar)] attribute, returning “bar”

get_struct_fields 🔒

parse_int_expr 🔒